Add room page and secure admin
