+++ app/admin/page.tsx
@@ -1,50 +1,11 @@
-'use client';
-import useSWR from 'swr'
-import { Card, CardContent, CardFooter, CardHeader, CardTitle } from '@/components/ui/card'
-import { Button } from '@/components/ui/button'
-
-const fetcher = (url: string) => fetch(url).then(res => res.json());
+import { cookies } from 'next/headers';
+import { redirect } from 'next/navigation';
+import AdminClient from './AdminClient';
 
 export default function AdminPage() {
-  const { data, mutate } = useSWR('/api/bookings', fetcher);
-
-  const createRoom = async (id: string) => {
-    const res = await fetch('/api/create-room', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ id })
-    });
-    const { url } = await res.json();
-    alert(`Room created: ${url}`);
-    mutate();
-  };
-
-  if (!data) return <p>Loading...</p>
-
-  return (
-    <main className="container mx-auto p-4">
-      <Card>
-        <CardHeader>
-          <CardTitle>Bookings</CardTitle>
-        </CardHeader>
-        <CardContent className="space-y-4">
-          {data.map((b: any) => (
-            <Card key={b.id} className="border p-4">
-              <div className="font-medium">
-                {b.date} {b.time}
-              </div>
-              <div className="text-sm text-muted-foreground mb-2">{b.notes}</div>
-              {!b.roomUrl ? (
-                <Button size="sm" onClick={() => createRoom(b.id)}>
-                  Create Room
-                </Button>
-              ) : (
-                <div className="text-sm break-all">Room: {b.roomUrl}</div>
-              )}
-            </Card>
-          ))}
-        </CardContent>
-      </Card>
-    </main>
-  )
+  const cookie = cookies().get('auth');
+  if (!cookie) {
+    redirect('/login');
+  }
+  return <AdminClient />;
 }
